requirement.txt
wikipedia-api
transformers
sentence-transformers
faiss-cpu
numpy
